requirements.txt
altair==5.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.8.0b4
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scipy==1.13.1
seaborn==0.13.2
setuptools==71.0.4
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.36.0
sympy==1.13.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.64
ultralytics-thop==2.0.0
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.1
wcwidth==0.2.13
wheel==0.43.0
widgetsnbextension==4.0.11
win-inet-pton==1.1.0
zipp==3.19.2
zstandard==0.23.0
